Updated to unity version 2018.2.14f1 among other things.
The update caused a lot of scripts to be modified by the unity editor. Have to check these changes soon.
Made the menu load asynchronously while the introduction is playing.
Made an 'oof' noise play when the character smashes into a wall (not fully functional)
The lighting data broke in lots all scenes. As such I re-generated the lighting on the menu, intro, lobby and the desert stage.
  The update seems to have made the lighting in the lobby go really wonky, so I changed some stuff to make it look better. It is not perfect but it will do for now.

Known issue:
  Pausing the game causes the music to rapidly switch (might be in the original idk)